Move to internal logging API
