+++ src/main/java/io/vertx/core/impl/CloseHooks.java
@@ -15,7 +15,7 @@ import io.vertx.core.AsyncResult;
 import io.vertx.core.Closeable;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
-import io.vertx.core.logging.Logger;
+import io.vertx.core.impl.logging.Logger;
 
 import java.util.HashSet;
 import java.util.Set;
