+++ src/main/java/io/vertx/core/impl/verticle/CompilingClassLoader.java
@@ -11,8 +11,8 @@
 
 package io.vertx.core.impl.verticle;
 
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
+import io.vertx.core.impl.logging.Logger;
+import io.vertx.core.impl.logging.LoggerFactory;
 
 import javax.tools.Diagnostic;
 import javax.tools.DiagnosticCollector;
