+++ src/main/java/io/vertx/core/net/impl/NetSocketImpl.java
@@ -31,8 +31,8 @@ import io.vertx.core.eventbus.MessageConsumer;
 import io.vertx.core.impl.ContextInternal;
 import io.vertx.core.impl.NetSocketInternal;
 import io.vertx.core.impl.VertxInternal;
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
+import io.vertx.core.impl.logging.Logger;
+import io.vertx.core.impl.logging.LoggerFactory;
 import io.vertx.core.net.NetSocket;
 import io.vertx.core.net.SocketAddress;
 import io.vertx.core.spi.metrics.TCPMetrics;
