+++ src/main/java/io/vertx/core/spi/resolver/ResolverProvider.java
@@ -19,8 +19,8 @@ import io.vertx.core.dns.AddressResolverOptions;
 import io.vertx.core.impl.VertxImpl;
 import io.vertx.core.impl.resolver.DnsResolverProvider;
 import io.vertx.core.impl.resolver.DefaultResolverProvider;
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
+import io.vertx.core.impl.logging.Logger;
+import io.vertx.core.impl.logging.LoggerFactory;
 
 import java.net.InetSocketAddress;
