+++ src/test/java/io/vertx/core/net/ProxyErrorTest.java
@@ -23,8 +23,6 @@ import io.vertx.core.VertxOptions;
 import io.vertx.core.http.HttpClient;
 import io.vertx.core.http.HttpClientOptions;
 import io.vertx.core.http.HttpClientResponse;
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
 import io.vertx.test.fakedns.FakeDNSServer;
 
 /**
